% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthdy.R
\name{birthday}
\alias{birthday}
\title{Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{Integer or numeric vector. The number of people in the group (sample size)}
}
\description{
Computes the probability that at least two people in a group share the same birthday.
}
\examples{
birthday(20:24)

}
