% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andersonsample.R
\name{mysample}
\alias{mysample}
\title{andersonsample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Integer. Sample size.}

\item{iter}{Integer. Number of iterations. Default is 10.}

\item{time}{Numeric. Time parameter. Default is 0.5.}
}
\value{
Output of the sampling function.
}
\description{
andersonsample
}
