% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on plane}

\item{gamma}{probability of overbooking}

\item{p}{probability passenger will show}
}
\value{
list of nc, nd, N, p, gamma, and a plot
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N = 400, gamma = 0.02, p = 0.95)}
}
